Added group column in evaluation list.
Realized by asking registration information in the GQL request, and by building a dict of learner=>group before building the evaluation list.  The group info is added for each learner based on this dict.  Learners without group are handled.

Tested manually in two scenarii : one with two students and differents groups, one with a student with a group and one without.